ReturnItemDraft to interface. #942
