+++ sphere-models/src/main/java/io/sphere/sdk/orders/ReturnItemDraft.java
@@ -1,45 +1,25 @@
 package io.sphere.sdk.orders;
 
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import javax.annotation.Nullable;
 
-public class ReturnItemDraft extends Base {
-    private final Long quantity;
-    private final String lineItemId;
-    private final ReturnShipmentState shipmentState;
+@JsonDeserialize(as = ReturnItemDraftImpl.class)
+public interface ReturnItemDraft {
+    Long getQuantity();
+
+    String getLineItemId();
+
+    ReturnShipmentState getShipmentState();
+
     @Nullable
-    private final String comment;
+    String getComment();
 
-    private ReturnItemDraft(final Long quantity, final String lineItemId, final ReturnShipmentState shipmentState, @Nullable final String comment) {
-        this.quantity = quantity;
-        this.lineItemId = lineItemId;
-        this.shipmentState = shipmentState;
-        this.comment = comment;
+    static ReturnItemDraft of(final long quantity, final String lineItemId, final ReturnShipmentState shipmentState, @Nullable final String comment) {
+        return new ReturnItemDraftImpl(quantity, lineItemId, shipmentState, comment);
     }
 
-    public static ReturnItemDraft of(final long quantity, final String lineItemId, final ReturnShipmentState shipmentState, @Nullable final String comment) {
-        return new ReturnItemDraft(quantity, lineItemId, shipmentState, comment);
-    }
-
-    public static ReturnItemDraft of(final long quantity, final String lineItemId, final ReturnShipmentState shipmentState) {
+    static ReturnItemDraft of(final long quantity, final String lineItemId, final ReturnShipmentState shipmentState) {
         return of(quantity, lineItemId, shipmentState, null);
-    }
-
-    public Long getQuantity() {
-        return quantity;
-    }
-
-    public String getLineItemId() {
-        return lineItemId;
-    }
-
-    public ReturnShipmentState getShipmentState() {
-        return shipmentState;
-    }
-
-    @Nullable
-    public String getComment() {
-        return comment;
     }
 }
